dev: fix fork bug
