+++ kotishell/src/execute/execute_a.c
@@ -6,7 +6,7 @@
 /*   By: elenavoronin <elnvoronin@gmail.com>          +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/07/25 15:48:11 by elenavoroni   #+#    #+#                 */
-/*   Updated: 2023/08/09 16:17:11 by dkolodze      ########   odam.nl         */
+/*   Updated: 2023/08/09 21:52:31 by dkolodze      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ int	ex_execute(t_ks_kotistate *state, t_li_node *cmds)
 	if (plumbing.fildesarr == NULL)
 		return (88);
 	l_ex_launch_all(state, cmds, plumbing);
-	retcode = l_ex_wait_all(cmds);
 	l_ex_full_plumbing_cleanup(&plumbing);
+	retcode = l_ex_wait_all(cmds);
 	return (retcode);
 }
