+++ test/ex_tests/.gitignore
@@ -1,3 +1,3 @@
 *.o
 objs
-ex_tests
+*.exe
