+++ test/ex_tests/Makefile
@@ -6,11 +6,11 @@
 #    By: elenavoronin <elnvoronin@gmail.com>          +#+                      #
 #                                                    +#+                       #
 #    Created: 2023/06/30 19:02:22 by evoronin      #+#    #+#                  #
-#    Updated: 2023/08/02 16:19:03 by dkolodze      ########   odam.nl          #
+#    Updated: 2023/08/09 21:57:09 by dkolodze      ########   odam.nl          #
 #                                                                              #
 # **************************************************************************** #
 
-NAME = ex_tests
+NAME = ex_test_ls.exe ex_test_ls_grep.exe ex_test_string_part.exe
 CC = cc
 CFLAGS = -Wall -Wextra -Werror -I ../../kotishell/include -g
 SRCS_FOLDER = ../../kotishell/src
@@ -32,27 +32,25 @@ execute/execute_d.c \
 execute/execute_e.c \
 execute/execute_f.c
 
-TEST_SRCS_NAMES =  ex_main.c
+TEST_SRCS_NAMES =  ex_test_ls.c ex_test_ls_grep.c ex_test_string_part.c
 TEST_OBJS = $(addprefix $(OBJS_DIR)/,$(TEST_SRCS_NAMES:.c=.o))
 OBJS_DIR = ./objs
 ABSOLUTE_SRCS_NAMES = $(addprefix $(SRCS_FOLDER)/,$(SRCS_NAMES))
 ABSOLUTE_OBJS = $(addprefix $(OBJS_DIR)/,$(SRCS_NAMES:.c=.o))
 
-all:  $(NAME)
+all: $(NAME)
 
 $(ABSOLUTE_OBJS): $(OBJS_DIR)/%.o: $(SRCS_FOLDER)/%.c
-	@mkdir -p $(@D)
-	@$(CC) $(CFLAGS) -c $^ -o $@
+	mkdir -p $(@D)
+	$(CC) $(CFLAGS) -c $^ -o $@
 
 $(TEST_OBJS): $(OBJS_DIR)/%.o: %.c
-	@mkdir -p $(@D)
-	@$(CC) $(CFLAGS) -c $^ -o $@
+	mkdir -p $(@D)
+	$(CC) $(CFLAGS) -c $^ -o $@
 
-OBJS=$(TEST_OBJS) $(ABSOLUTE_OBJS)
-
-$(NAME): $(OBJS)
-	echo $(OBJS)
-	@$(CC) $(CFLAGS) $(OBJS) -o $(NAME)
+$(NAME): %.exe: %.o $(ABSOLUTE_OBJS)
+	echo $^
+	$(CC) $(CFLAGS) $^ -o $@
 
 clean:
 	rm -rf $(OBJS_DIR)
